tests: Split test_message_edit.py.
